Remove general parts of scenario builder
